Modify tests for faker data

+++ spec/requests/api/v1/merchants_request_spec.rb
@@ -34,14 +34,13 @@ describe "Merchant API" do
         expect(response).to be_successful
         merchant_items = JSON.parse(response.body, symbolize_names: true)
         expect(merchant_items[:data].count).to eq(4)
-        
         merchant_items[:data].each do |item|
             expect(item[:attributes]).to have_key(:name)
             expect(item[:attributes][:name]).to be_a(String)
             expect(item[:attributes]).to have_key(:description)
             expect(item[:attributes][:description]).to be_a(String)
             expect(item[:attributes]).to have_key(:unit_price)
-            expect(item[:attributes][:unit_price]).to be_a(Float)
+            expect(item[:attributes][:unit_price]).to be_a(Integer)
             expect(item[:attributes]).to have_key(:merchant_id)
             expect(item[:attributes][:merchant_id]).to be_a(Integer)
             expect(Merchant.find(item[:attributes][:merchant_id])).to be_a(Merchant)
